requirements.txt
opencv-python-headless
numpy
requests
pytesseract
Pillow
pyyaml
scikit-image
